requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-jose[cryptography]